E-13: extract AssemblyAttributeEmitter from ReflectionMetadataEmitter (#1361)
Moves the assembly-attribute band (EmitReferenceAssemblyAttribute,
ParseAssemblyVersion, EmitAssemblyInteropAttributes,
EmitUserAssemblyAttributes, EmitFriendAssemblyAttributes,
EmitGSharpTypeSemantics, EmitDebuggableAttribute,
EmitNullableContextAttribute — RME lines 3599-3885) verbatim into a new
internal sealed AssemblyAttributeEmitter.

Wiring (ADR-0150 composition idiom, 5 deps — below the delegate
ceiling, so no back-reference):
  ctor(EmitContext, MetadataTokenCache, WellKnownReferences,
       CustomAttributeEncoder, Func<Type, TypeReferenceHandle>)
constructed in EmitCore immediately after customAttrEncoder (it
forwards blob writes into the encoder and needs wellKnown /
GetTypeReference, which close over Core* types materialised in
EmitCore). The only mechanical rewrites are this.GetTypeReference ->
this.getTypeReference (x2), this.customAttrEncoder. ->
this.attrEncoder. (x4), and public visibility on the four EmitCore
entry points (EmitReferenceAssemblyAttribute, ParseAssemblyVersion,
EmitAssemblyInteropAttributes, EmitDebuggableAttribute); the other
four methods stay private to the new class.

EmitCore call sites re-pointed to assemblyAttrs.*; no other callers
exist anywhere in the Emit folder (or repo), so no transitional
forwarders were needed. Stale "orchestrators stay on the root"
comments in CustomAttributeEncoder / RME / ImportedAssemblySemantics
updated to match.

Verified: Release build 0 warnings / 0 errors; 119 samples/*.gs
compiled before/after with the Release gsc and all emitted PEs are
byte-identical (3 samples fail at baseline for unrelated reasons:
GsharpExtensionsMixed/Optional/Sequences need Gsharp.Extensions refs).
